# shared/openapi.yaml
openapi: 3.1.0
info:
  title: EasySSH API
  description: |
    EasySSH 后端 API 规范 - 现代化的 SSH 管理平台

    ## 功能模块
    - 用户认证与授权
    - 服务器管理（CRUD）
    - SSH 连接与终端
    - SFTP 文件传输
    - 批量操作与脚本执行
    - 系统监控与日志

  version: 1.0.0
  contact:
    name: EasySSH Team
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8521/api/v1
    description: 本地开发环境
  - url: https://api.easyssh.example.com/api/v1
    description: 生产环境
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 访问令牌
...
paths:
...
  /ai/config:
    get:
      tags:
        - ai
      summary: 获取 AI 配置状态
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 返回当前用户可用的 AI 配置状态
          content:
            application/json:
              schema:
                type: object
                properties:
                  configured:
                    type: boolean
                  provider:
                    type: string
                  model:
                    type: string
                  models:
                    type: array
                    items:
                      type: string
                  has_key:
                    type: boolean
                  message:
                    type: string
...